new lint: match_like_matches_macro
